refactor: create error sub package
